Remove Environment trait.
